Implement rest api populate syntax and Init  refactor convert query params for v4
