+++ packages/core/utils/lib/index.js
@@ -4,11 +4,7 @@
  * Export shared utilities
  */
 const { buildQuery, hasDeepFilters } = require('./build-query');
-const {
-  convertRestQueryParams,
-  VALID_REST_OPERATORS,
-  QUERY_OPERATORS,
-} = require('./convert-rest-query-params');
+const { VALID_REST_OPERATORS, QUERY_OPERATORS } = require('./convert-query-params');
 const parseMultipartData = require('./parse-multipart');
 const sanitizeEntity = require('./sanitize-entity');
 const parseType = require('./parse-type');
@@ -43,7 +39,6 @@ module.exports = {
   finder,
   policy,
   templateConfiguration,
-  convertRestQueryParams,
   VALID_REST_OPERATORS,
   QUERY_OPERATORS,
   buildQuery,
